Fix: Evaluate() prep work
In an effort to not use eval(), even if properly secured, and also to have a properly functioning basic calculator akin to the MATE one, and without using algorithms that for now are too complicated for me to follow, I have determined that the best course of action would probably be to simply use math.js. So I will come back to this at a later point to implement the package properly.